Fixed slope collision for the most part. At this point, most of the remaining issues with slopes should be issues with other parts of movement, not the slopes themselves.
Fixed some alpha BlendMode nonsense. The AlphaBlend mode treats everything as premultiplied, but I wasn't multiplying EVERY color channel by the opacity when trying to make things translucent, just the alpha.

Added a debug function to track Kirby's position history. Currently hold left Alt to record, and press left Control to clear history.

+++ Constants.cs
@@ -63,9 +63,9 @@ namespace KirbyNightmareInDreamLand
             public const int GENTLE1_SLOPE_LEFT_YINTERCEPT = 0;
 
             public const float GENTLE2_SLOPE_LEFT_M = 0.5F;
-            public const float GENTLE2_SLOPE_LEFT_YINTERCEPT = 8F;
+            public const int GENTLE2_SLOPE_LEFT_YINTERCEPT = 8;
 
-            public const int STEEP_SLOPE_LEFT_M = 1;
+            public const float STEEP_SLOPE_LEFT_M = 1;
             public const int STEEP_SLOPE_LEFT_YINTERCEPT = 0;
 
             public const float GENTLE1_SLOPE_RIGHT_M = -0.5F;
@@ -74,7 +74,7 @@ namespace KirbyNightmareInDreamLand
             public const float GENTLE2_SLOPE_RIGHT_M = -0.5F;
             public const int GENTLE2_SLOPE_RIGHT_YINTERCEPT = 16;
 
-            public const int STEEP_SLOPE_RIGHT_M = -1;
+            public const float STEEP_SLOPE_RIGHT_M = -1;
             public const int STEEP_SLOPE_RIGHT_YINTERCEPT = 16;
         }
         public static class HitBoxes
